Add files via upload
Now, when the user interrupts the execution, the timer for the relay restarts. This is because of a biological reason. All the research made when starting this show that for sterilization a single dosage of radiation is applied, meaning that the period of exposure must be one from beginning to end. So, if the user opens the door, it should restart the process.